+++ docker-compose.yml
@@ -5,7 +5,7 @@ services:
       dockerfile: Dockerfile
     container_name: wb-backend-database
     ports:
-      - "5432:5432"
+      - "7777:5432"
     environment:
       - POSTGRES_USER=postgres
       - POSTGRES_PASSWORD=postgres
