+++ host-electron/.eslintrc.cjs
@@ -5,5 +5,8 @@ module.exports = {
     'plugin:react/jsx-runtime',
     '@electron-toolkit/eslint-config-ts/recommended',
     '@electron-toolkit/eslint-config-prettier'
-  ]
+  ],
+  "rules": {
+    "prettier/prettier": "off"
+  }
 }
