+++ host-electron/src/renderer/src/Views/GalleryLab.tsx
@@ -1,8 +1,8 @@
 import { PiRocketLaunchThin } from 'react-icons/pi'
-import SideComponent from '../components/Components/SideComponent.tsx'
+import SideComponent from '../components/SideComponent.tsx'
 import React, { useEffect, useState } from 'react'
-import GalleryInput from '../components/Components/GalleryInput'
-import ModalComponent from '../components/Components/ModalComponent.tsx'
+import GalleryInput from '../components/GalleryInput'
+import ModalComponent from '../components/ModalComponent.tsx'
 import { useLocation } from 'react-router-dom'
 import { useNavigate } from 'react-router-dom'
 import axios from 'axios'
