+++ host-electron/src/renderer/src/components/CarouselComponent.tsx
@@ -1,0 +1,33 @@
+import Carousel from "react-bootstrap/Carousel";
+
+import ExampleCarouselImage from "./ExampleCarouselImage";
+import ExampleCarouselImage2 from "./ExampleCarouselImage2";
+import ExampleCarouselImage3 from "./ExampleCarouselImage3";
+
+export default function CarouselComponent() {
+  return (
+    <Carousel>
+      <Carousel.Item>
+        <ExampleCarouselImage text="First slide" />
+        <Carousel.Caption>
+          {/* <h3>First slide label</h3>
+          <p>Nulla vitae elit libero, a pharetra augue mollis interdum.</p> */}
+        </Carousel.Caption>
+      </Carousel.Item>
+      <Carousel.Item>
+      <ExampleCarouselImage2 text="Second Slide"/>
+        <Carousel.Caption>
+          
+        </Carousel.Caption>
+      </Carousel.Item>
+      <Carousel.Item>
+        <ExampleCarouselImage3 text="Third slide" />
+        <Carousel.Caption>
+          <h3>Third slide label</h3>
+          <p>Praesent commodo cursus magna, vel scelerisque nisl consectetur.</p>
+        </Carousel.Caption>
+      </Carousel.Item>
+    </Carousel>
+  );
+}
+
