--- host-electron/src/renderer/src/components/Components/ExampleCarouselImage2.tsx
@@ -1,10 +1,0 @@
-import arraaaaam from "/src/assets/meme.webp"
-
-export default function ExampleCarouselImage2(props: { text: string }) {
-    return (
-      <img
-        src={arraaaaam}
-        className="d-block w-100"
-      />
-    );
-  }
