--- host-electron/src/renderer/src/components/Components/ExampleCarouselImage3.tsx
@@ -1,10 +1,0 @@
-import durag from "/src/assets/durag.webp"
-
-export default function ExampleCarouselImage3(props: { text: string }) {
-    return (
-      <img
-        src={durag}
-        className="d-block w-100"
-      />
-    );
-  }
