--- host-electron/src/renderer/src/components/Components/PrimaryNavigationBar.tsx
@@ -1,7 +1,0 @@
-export default function PrimaryNavigationBar(){
-    return(
-        <div>
-            
-        </div>
-    )
-}
