--- host-electron/src/renderer/src/components/Components/SplitFrame.tsx
@@ -1,10 +1,0 @@
-export default function SplitFrame() {
-  return (
-    <div>
-      <div>
-        <div></div>
-      </div>
-      <div></div>
-    </div>
-  );
-}
