+++ host-electron/src/renderer/src/components/EmotiBitList.tsx
