+++ host-electron/src/renderer/src/components/ExampleCarouselImage2.tsx
@@ -1,0 +1,10 @@
+import arraaaaam from "/src/assets/meme.webp"
+
+export default function ExampleCarouselImage2(props: { text: string }) {
+    return (
+      <img
+        src={arraaaaam}
+        className="d-block w-100"
+      />
+    );
+  }
