+++ host-electron/src/renderer/src/components/ExampleCarouselImage3.tsx
@@ -1,0 +1,10 @@
+import durag from "/src/assets/durag.webp"
+
+export default function ExampleCarouselImage3(props: { text: string }) {
+    return (
+      <img
+        src={durag}
+        className="d-block w-100"
+      />
+    );
+  }
