+++ host-electron/src/renderer/src/components/GalleryInput.tsx
