+++ host-electron/src/renderer/src/components/PrimaryNavigationBar.tsx
@@ -1,0 +1,7 @@
+export default function PrimaryNavigationBar(){
+    return(
+        <div>
+            
+        </div>
+    )
+}
