+++ host-electron/src/renderer/src/components/SplitFrame.tsx
@@ -1,0 +1,10 @@
+export default function SplitFrame() {
+  return (
+    <div>
+      <div>
+        <div></div>
+      </div>
+      <div></div>
+    </div>
+  );
+}
